requirements.txt
google-api-python-client~=2.135.0
google-auth
google-auth-oauthlib
google-auth-httplib2
mediapipe~=0.10.11
opencv-python~=4.10.0.84
gspread~=6.1.2
numpy~=1.24.4
oauth2client~=4.1.3